+++ packages/core/src/index.ts
@@ -1,5 +1,6 @@
 export * from './adapter.js';
 export * from './executor.js';
 export * from './length.js';
+export * from './measure.js';
 export * from './not-implemented.js';
 export * from './schema.js';