+++ packages/workspace/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @deckflow/deckuse-workspace
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies [6d11f42]
+  - @deckflow/deckuse-core@1.2.0
+
 ## 1.1.0
 
 ### Minor Changes